+++ CHANGELOG.md
@@ -26,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `StrokeOptions::upx_wide` returns stroke options for a given unsigned pixel
   stroke width.
 - `LazyTexture` now implements `PartialEq`.
+- `WindowBehavior::unrecoverable_error` is a new function that allows
+  applications to take control when an unrecoverable error occurs. Previously,
+  these errors would panic, and the provided implementation is a panic.
 
 ## v0.11.0 (2024-09-14)